feat(orm): partial drizzle postgres setup

+++ src/helpers/setup-payload.ts
@@ -2,15 +2,21 @@ import consola from 'consola';
 import { execa } from 'execa';
 import fs from 'fs-extra';
 
-export default async function setupPayload(projectPath: string) {
+export default async function setupPayload(projectPath: string | undefined) {
+  if (!projectPath) {
+    consola.error('Project path is not defined');
+    process.exit(1);
+  }
   try {
     consola.info('Setting up payload');
     Promise.all([
       await fs.remove(`${projectPath}/app/layout.tsx`),
-      await execa`npx create-payload-app@beta`
-    ])
+      await execa({ timeout: 5000 })`npx create-payload-app@beta`,
+    ]);
     consola.success('Payload');
+    process.exit(0);
   } catch (error) {
     consola.error('Error setting up payload');
+    process.exit(1);
   }
 }
